+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `save_special_prices`: add a type hint and document the return value
 * `delete_special_prices`: add a type hint for the return value
+* Add the `TypedDict` `BasePrice`
 
 ## 2.1.0 (2024/11/27)